requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.2
openai==1.95.0
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
XlsxWriter==3.2.2
